requirements.txt
# Data generation
openai
pandas
python-dotenv

# Model training
torch
transformers
scikit-learn
tqdm
sentencepiece
